// src/generated/models/MicrosoftTo_Do_Business_Model.ts
export type ToDo_V2importance = 'low'|'normal'|'high';
export type ToDo_V2status = 'notStarted'|'inProgress'|'completed'|'waitingOnOthers'|'deferred';

export interface ToDo_V2 {
  "@odata.context"?: string;
  "@odata.etag"?: string;
  // Unique identifier of the to-do.
  id?: string;
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  createdDateTime?: string;
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  lastModifiedDateTime?: string;
  // To-do body that typically contains information about the to-do.
  body?: {
    // The type of the content. Possible values are text and html.
    contentType?: string;
    // The content of the item.
    content?: string;
  };
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  bodyLastModifiedDateTime?: string;
  // The date in the specified time zone that the task was finished.
  completedDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // The date in the specified time zone that the task is to be finished.
  dueDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // Low, normal or high.
  importance?: ToDo_V2importance;
  // True if an alert is set to remind the user of the to-do.
  isReminderOn?: boolean;
  // The date and time for a reminder alert of the task to occur.
  reminderDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // Indicates state or progress of the to-do - not started, in progress, completed, waiting on others or deferred.
  status?: ToDo_V2status;
  // Brief description or title of the to-do.
  title?: string;
}